Continue MongoDB Migration

+++ couchbase/dbtypes.go
@@ -2,9 +2,9 @@ package couchbase
 
 // Data format of the CouchBase db
 type CouchData struct {
-	BuildingName  string `json:"buildingName"`
-	EnergyType    string `json:"energyType"`
-	EnergyUnit    string `json:"energyUnit"`
-	EnergyValue   int    `json:"energyValue"`
-	UnixTimeValue int    `json:"unixTimeValue"`
+	BuildingName  string  `json:"buildingName"`
+	EnergyType    string  `json:"energyType"`
+	EnergyUnit    string  `json:"energyUnit"`
+	EnergyValue   float64 `json:"energyValue"`
+	UnixTimeValue int     `json:"unixTimeValue"`
 }
